+++ src/layout/test/test-utils.js
@@ -1,0 +1,14 @@
+var assert = require('chai').assert;
+
+function deepApproximatelyEqual(t, o, precision) {
+  for (var k in t) {
+    if (typeof t[k] === 'object') {
+      deepApproximatelyEqual(t[k], o[k], precision);
+      continue;
+    }
+
+    assert.approximately(t[k], o[k], precision || 1e-7);
+  }
+}
+
+exports.deepApproximatelyEqual = deepApproximatelyEqual;