+++ src/layout/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 0.6.0
+
+- Moving `createGraphToViewportConversionFunction` to `conversion` submodule.
+
 ## 0.5.0
 
 - Adding `rotation`.
-- Adding `createGraphToViewportConversionFunction`.
+- Adding undocumented `createGraphToViewportConversionFunction`.
 - Random layout now accepts variable dimensions.
 - Circular layout now accepts custom dimensions.